all/examples/src/redis/requirements.txt
python-dotenv==0.16.0
redis==3.5.3
aioredis==2.0.0